+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+- Add clear all logs.
+
 ### Updated
 - Only allow admin show cp nav item and manage plugin settings.